CoreDataManaging protocol now a seperate file ready for new messaging model
